+++ Tests/RoktUXHelperTests/Services/Events/TestEventService.swift
@@ -208,6 +208,31 @@ final class TestEventService: XCTestCase {
         XCTAssertTrue(stubUXHelper.roktEvents.contains(.PlacementClosed))
     }
 
+    func test_sendDismissal_onInstantPurchaseDismissed_dismissalEventsAndSignals_shouldSend() throws {
+        // Arrange
+        let eventService = get_mock_event_processor(
+            startDate: startDate,
+            uxEventDelegate: stubUXHelper,
+            eventHandler: { event in
+                self.events.append(event)
+            })
+
+        // Act
+        eventService.dismissOption = .instantPurchaseDismiss
+        eventService.sendDismissalEvent()
+
+        // Assert
+        let event = events.first
+        XCTAssertEqual(event?.eventType, .SignalInstantPurchaseDismissal)
+        XCTAssertEqual(event?.pageInstanceGuid, mockPageInstanceGuid)
+        XCTAssertNotNil(event?.metadata.first { $0.name == kInitiator })
+        XCTAssertNotNil(event?.metadata.first { $0.value == kInstantPurchaseDismiss })
+
+        // Rokt callbacks
+        XCTAssertEqual(stubUXHelper.roktEvents.count, 1)
+        XCTAssertTrue(stubUXHelper.roktEvents.contains(.PlacementClosed))
+    }
+
     func test_dismissal_dimissed_event() throws {
         // Arrange
         let eventService = get_mock_event_processor(
@@ -316,119 +341,6 @@ final class TestEventService: XCTestCase {
         XCTAssertEqual(stubUXHelper.layoutId, "pluginId")
         XCTAssertEqual(stubUXHelper.url, "https://www.rokt.com")
         XCTAssertEqual(stubUXHelper.openUrlType, .passthrough)
-    }
-
-    func test_send_device_pay_initiated() {
-        // Arrange
-        let eventService = get_mock_event_processor(
-            startDate: startDate,
-            uxEventDelegate: stubUXHelper,
-            eventHandler: { event in
-                self.events.append(event)
-            })
-
-        // Act
-        eventService.cartItemDevicePay(catalogItem: .mock(), paymentProvider: .applePay)
-
-        // Assert
-        let event = events.first
-        XCTAssertEqual(event?.eventType, .SignalCartItemDevicePayInitiated)
-        XCTAssertEqual(event?.pageInstanceGuid, mockPageInstanceGuid)
-        let cartItemId = event?.eventData.first { $0.name == kCartItemId }
-        XCTAssertEqual(cartItemId?.value, "cartItemId")
-        let catalogItemId = event?.eventData.first { $0.name == kCatalogItemId }
-        XCTAssertEqual(catalogItemId?.value, "catalogItemId")
-        let currency = event?.eventData.first { $0.name == kCurrency }
-        XCTAssertEqual(currency?.value, "USD")
-        let description = event?.eventData.first { $0.name == kDescription }
-        XCTAssertEqual(description?.value, "Catalog Description")
-        let linkedProductId = event?.eventData.first { $0.name == kLinkedProductId }
-        XCTAssertEqual(linkedProductId?.value, "linked")
-        let totalPrice = event?.eventData.first { $0.name == kTotalPrice }
-        XCTAssertEqual(totalPrice?.value, "14.99")
-        let quantity = event?.eventData.first { $0.name == kQuantity }
-        XCTAssertEqual(quantity?.value, "1")
-        let unitPrice = event?.eventData.first { $0.name == kUnitPrice }
-        XCTAssertEqual(unitPrice?.value, "14.99")
-
-        // Rokt callbacks
-        XCTAssertEqual(stubUXHelper.roktEvents.count, 1)
-    }
-
-    func test_send_device_pay_succeeded() {
-        // Arrange
-        let eventService = get_mock_event_processor(
-            startDate: startDate,
-            catalogItems: [.mock(catalogItemId: "catalogItemId")],
-            uxEventDelegate: stubUXHelper,
-            eventHandler: { event in
-                self.events.append(event)
-            })
-
-        // Act
-        eventService.cartItemDevicePaySuccess(itemId: "catalogItemId")
-
-        // Assert
-        let event = events.first
-        XCTAssertEqual(event?.eventType, .SignalCartItemDevicePay)
-        XCTAssertEqual(event?.pageInstanceGuid, mockPageInstanceGuid)
-        let cartItemId = event?.eventData.first { $0.name == kCartItemId }
-        XCTAssertEqual(cartItemId?.value, "cartItemId")
-        let catalogItemId = event?.eventData.first { $0.name == kCatalogItemId }
-        XCTAssertEqual(catalogItemId?.value, "catalogItemId")
-        let currency = event?.eventData.first { $0.name == kCurrency }
-        XCTAssertEqual(currency?.value, "USD")
-        let description = event?.eventData.first { $0.name == kDescription }
-        XCTAssertEqual(description?.value, "Catalog Description")
-        let linkedProductId = event?.eventData.first { $0.name == kLinkedProductId }
-        XCTAssertEqual(linkedProductId?.value, "linked")
-        let totalPrice = event?.eventData.first { $0.name == kTotalPrice }
-        XCTAssertEqual(totalPrice?.value, "14.99")
-        let quantity = event?.eventData.first { $0.name == kQuantity }
-        XCTAssertEqual(quantity?.value, "1")
-        let unitPrice = event?.eventData.first { $0.name == kUnitPrice }
-        XCTAssertEqual(unitPrice?.value, "14.99")
-
-        // Rokt callbacks
-        XCTAssertEqual(stubUXHelper.roktEvents.count, 0)
-    }
-
-    func test_send_device_pay_failed() {
-        // Arrange
-        let eventService = get_mock_event_processor(
-            startDate: startDate,
-            catalogItems: [.mock(catalogItemId: "catalogItemId")],
-            uxEventDelegate: stubUXHelper,
-            eventHandler: { event in
-                self.events.append(event)
-            })
-
-        // Act
-        eventService.cartItemDevicePayFailure(itemId: "catalogItemId")
-
-        // Assert
-        let event = events.first
-        XCTAssertEqual(event?.eventType, .SignalCartItemDevicePayFailure)
-        XCTAssertEqual(event?.pageInstanceGuid, mockPageInstanceGuid)
-        let cartItemId = event?.eventData.first { $0.name == kCartItemId }
-        XCTAssertEqual(cartItemId?.value, "cartItemId")
-        let catalogItemId = event?.eventData.first { $0.name == kCatalogItemId }
-        XCTAssertEqual(catalogItemId?.value, "catalogItemId")
-        let currency = event?.eventData.first { $0.name == kCurrency }
-        XCTAssertEqual(currency?.value, "USD")
-        let description = event?.eventData.first { $0.name == kDescription }
-        XCTAssertEqual(description?.value, "Catalog Description")
-        let linkedProductId = event?.eventData.first { $0.name == kLinkedProductId }
-        XCTAssertEqual(linkedProductId?.value, "linked")
-        let totalPrice = event?.eventData.first { $0.name == kTotalPrice }
-        XCTAssertEqual(totalPrice?.value, "14.99")
-        let quantity = event?.eventData.first { $0.name == kQuantity }
-        XCTAssertEqual(quantity?.value, "1")
-        let unitPrice = event?.eventData.first { $0.name == kUnitPrice }
-        XCTAssertEqual(unitPrice?.value, "14.99")
-
-        // Rokt callbacks
-        XCTAssertEqual(stubUXHelper.roktEvents.count, 0)
     }
 
     func test_send_instant_purchase_initiated() {
@@ -642,9 +554,5 @@ class MockUXHelper: UXEventsDelegate {
 
     func onCartItemInstantPurchase(_ layoutId: String, catalogItem: RoktUXHelper.CatalogItem) {
         self.roktEvents.append(.CartItemInstantPurchase)
-    }
-
-    func onCartItemDevicePay(_ layoutId: String, catalogItem: RoktUXHelper.CatalogItem, paymentProvider: PaymentProvider) {
-        self.roktEvents.append(.CartItemDevicePay)
     }
 }