+++ Tests/RoktUXHelperTests/UI/ViewModel/Mocks/MockEventService.swift
@@ -22,9 +22,6 @@ class MockEventService: EventServicing {
     var cartItemInstantPurchaseCalled = false
     var cartItemInstantPurchaseSuccessCalled = false
     var cartItemInstantPurchaseFailureCalled = false
-    var cartItemDevicePayCalled = false
-    var cartItemDevicePaySuccessCalled = false
-    var cartItemDevicePayFailureCalled = false
 
     // MARK: - Protocol Properties
 
@@ -85,18 +82,6 @@ class MockEventService: EventServicing {
         cartItemInstantPurchaseFailureCalled = true
     }
 
-    func cartItemDevicePay(catalogItem: CatalogItem, paymentProvider: PaymentProvider) {
-        cartItemDevicePayCalled = true
-    }
-
-    func cartItemDevicePaySuccess(itemId: String) {
-        cartItemDevicePaySuccessCalled = true
-    }
-
-    func cartItemDevicePayFailure(itemId: String) {
-        cartItemDevicePayFailureCalled = true
-    }
-
     // MARK: - Additional Test Methods
 
     func sendImpressionEvents(currentOffer: Int) {
@@ -129,9 +114,6 @@ class MockEventService: EventServicing {
         cartItemInstantPurchaseCalled = false
         cartItemInstantPurchaseSuccessCalled = false
         cartItemInstantPurchaseFailureCalled = false
-        cartItemDevicePayCalled = false
-        cartItemDevicePaySuccessCalled = false
-        cartItemDevicePayFailureCalled = false
         dismissOption = nil
     }
 }